+++ src/features/system/index.ts
@@ -1,6 +1,8 @@
 export { PomodoroNotifyListener } from './PomodoroNotifyListener'
 export { PttListener } from './PttListener'
 export { QuitConfirmListener } from './QuitConfirmListener'
+export { WindowLayoutListener } from './WindowLayoutListener'
+export { resetWindowToDefault } from './windowLayout'
 export {
   useAutostart,
   type AutostartStatus,